# static/contract/openapi.json
{
  "components": {
    "schemas": {
      "BackupBody": {
        "properties": {
          "backup": {
            "title": "Backup",
            "type": "boolean"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "backup"
        ],
        "title": "BackupBody",
        "type": "object"
      },
      "CheckResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CheckedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "items"
        ],
        "title": "CheckResponse",
        "type": "object"
      },
...
      "FlagResponse": {
        "description": "Resposta de `/api/mode` e `/api/backup`: confirma e devolve a flag.",
        "properties": {
          "backup": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "zip_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip Only"
          }
        },
        "required": [
          "ok"
        ],
        "title": "FlagResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MapBody": {
        "properties": {
          "branch": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name",
          "url"
        ],
        "title": "MapBody",
        "type": "object"
      },
      "ModeBody": {
        "properties": {
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "zip_only": {
            "title": "Zip Only",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "zip_only"
        ],
        "title": "ModeBody",
        "type": "object"
      },
      "OkResponse": {
        "description": "Resposta das rotas que só confirmam: `/api/map`.",
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
      "RollbackResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "restored_from": {
            "title": "Restored From",
            "type": "string"
          },
          "zip_only": {
            "title": "Zip Only",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "name",
          "restored_from",
          "path",
          "zip_only"
        ],
        "title": "RollbackResponse",
        "type": "object"
      },
      "ScanResponse": {
        "properties": {
          "backup": {
            "title": "Backup",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScannedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "zip_only": {
            "title": "Zip Only",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "items",
          "zip_only",
          "backup"
        ],
        "title": "ScanResponse",
        "type": "object"
      },
...
      "SuggestResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SuggestResponse",
        "type": "object"
      },
...
      "UpdateBody": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name"
        ],
        "title": "UpdateBody",
        "type": "object"
      },
      "UpdateResponse": {
        "properties": {
          "backup": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup"
          },
          "download_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Bytes"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "new_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Bytes"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "old_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Bytes"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "remote_sha": {
            "title": "Remote Sha",
            "type": "string"
          },
          "speed_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Bps"
          },
          "zip_only": {
            "title": "Zip Only",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "name",
          "remote_sha",
          "zip_only",
          "path"
        ],
        "title": "UpdateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RepoRefresh",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backup": {
      "post": {
        "operationId": "api_backup_api_backup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Backup"
      }
    },
    "/api/check": {
      "get": {
        "operationId": "api_check_api_check_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Check"
      }
    },
...
    "/api/map": {
      "post": {
        "operationId": "api_map_api_map_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Map"
      }
    },
    "/api/mode": {
      "post": {
        "operationId": "api_mode_api_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Mode"
      }
    },
    "/api/rollback": {
      "post": {
        "operationId": "api_rollback_api_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Rollback"
      }
    },
    "/api/scan": {
      "get": {
        "operationId": "api_scan_api_scan_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Scan"
      }
    },
    "/api/suggest": {
      "get": {
        "operationId": "api_suggest_api_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Suggest"
      }
    },
...
    "/api/update": {
      "post": {
        "operationId": "api_update_api_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update"
      }
    }
  }
}